+++ .prettierrc
@@ -1,5 +1,8 @@
 {
+  "semi": true,
+  "singleQuote": true,
   "trailingComma": "es5",
+  "bracketSpacing": true,
   "printWidth": 120,
   "arrowParens": "always"
 }
