+++ demo/package.json
@@ -15,8 +15,10 @@
   "homepage": "https://antvis.github.io/color-palette-json-schema",
   "dependencies": {
     "@ant-design/icons": "^4.6.2",
+    "@types/chroma-js": "^2.1.3",
     "ajv": "^8.6.1",
     "antd": "^4.16.3",
+    "classnames": "^2.3.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
@@ -27,12 +29,10 @@
     "@babel/preset-env": "^7.14.7",
     "@babel/preset-react": "^7.14.5",
     "@babel/preset-typescript": "^7.14.5",
-    "@types/chroma-js": "^2.1.3",
     "@types/react": "^17.0.11",
     "@types/react-dom": "^17.0.8",
     "babel-loader": "^8.2.2",
     "babel-plugin-import": "^1.13.3",
-    "classnames": "^2.3.1",
     "css-loader": "^5.2.6",
     "file-loader": "^6.2.0",
     "gh-pages": "^3.2.3",
