+++ demo/src/app.tsx
@@ -1,24 +1,13 @@
 import React, { useState } from 'react';
 import Ajv from 'ajv';
-import chroma from 'chroma-js';
 import { Button, Dropdown, Menu } from 'antd';
 import { DownOutlined } from '@ant-design/icons';
-import colorSchema from '../build/color-schema.json';
-import {
-  Color,
-  ContinuousColor,
-  Palette,
-  CategoricalPalette as CatP,
-  DiscreteScalePalette as DisP,
-  ContinuousScalePalette as ConP,
-  MatrixPalette as MatP,
-  ColorSchema,
- colorToHex } from '../src';
-import classic from '../examples/classic.json';
-import * as antdColors from '../examples/antd-color';
-import antvColor from '../examples/antv-color.json';
+import colorSchema from '../../build/color-schema.json';
+import { Palette, colorToHex } from '../../src';
+import classic from '../../examples/classic.json';
+import * as antdColors from '../../examples/antd-color';
+import antvColor from '../../examples/antv-color.json';
 import Swatch from './components/Swatch';
-
 
 const examples: Record<string, any> = {
   'Classic Demo': classic,
@@ -30,113 +19,6 @@ const examples: Record<string, any> = {
 
 const ajv = new Ajv();
 const validateSchema = ajv.compile(colorSchema);
-
-const ColorCell = (props: { color: Color }) => {
-  const { color } = props;
-  const tooltip = Object.keys(color)
-    .map((key) => `${key}: ${color[key as keyof Color]}`)
-    .join('\n');
-  const fill = colorToHex(color);
-  return (
-    <div className="color-cell">
-      <svg>
-        <rect width="20" height="20" rx="4" ry="4" fill={fill}>
-          <title>{tooltip}</title>
-        </rect>
-      </svg>
-    </div>
-  );
-};
-
-const CategoricalPalette = (props: { palette: CatP }) => {
-  const { palette } = props;
-  return (
-    <div className="palette">
-      {palette.colors.map((color, index) => (
-        <ColorCell key={index} color={color} />
-      ))}
-    </div>
-  );
-};
-
-const DiscretePalette = (props: { palette: DisP }) => {
-  const { palette } = props;
-  const { colors } = palette;
-  const len = colors.length;
-  const fwArray = [...Array(100)];
-  const colorFunc = (per: number) => {
-    const i = Math.floor(per / (1 / len));
-    const color = colors[i];
-    return colorToHex(color);
-  };
-
-  return (
-    <div className="palette">
-      <div className="gradient">
-        {fwArray.map((_, index) => {
-          return (
-            <span className="grad-step" key={index} style={{ backgroundColor: `${colorFunc(index / 100)}` }}></span>
-          );
-        })}
-      </div>
-    </div>
-  );
-};
-
-const colorsHaveLocation = (colors: ContinuousColor[]): boolean => {
-  for (const color of colors) {
-    if (typeof color.location === 'undefined') {
-      return false;
-    }
-  }
-  return true;
-};
-
-const ContinuousPalette = (props: { palette: ConP }) => {
-  const { palette } = props;
-  const { colors } = palette;
-  const fwArray = [...Array(100)];
-  const haveLocation = colorsHaveLocation(colors);
-  let domain = [0, 1];
-  if (haveLocation) {
-    colors.sort((color1, color2) => (color1.location || 0) - (color2.location || 0));
-    domain = colors.map((color) => color.location || 0);
-  }
-  const colorScaleFunc = chroma.scale(colors.map((color) => colorToHex(color))).domain(domain);
-
-  return (
-    <div className="palette">
-      <div className="gradient">
-        {fwArray.map((_, index) => {
-          return (
-            <span
-              className="grad-step"
-              key={index}
-              style={{
-                backgroundColor: `${colorScaleFunc(index / 100).hex()}`,
-              }}
-            ></span>
-          );
-        })}
-      </div>
-    </div>
-  );
-};
-
-const MatrixPalette = (props: { palette: MatP }) => {
-  const { palette } = props;
-  return (
-    <div className="matrixpalette">
-      {palette.colors.map((colors, index1) => (
-        <div key={index1} className="matrixPalette-row">
-          {colors.map((color, index2) => (
-            <ColorCell key={index2} color={color} />
-          ))}
-        </div>
-      ))}
-    </div>
-  );
-};
 
 const ColorPaletteView = (props: { palette: Palette }) => {
   const { palette } = props;
