+++ demo/src/components/MiniSwatch/index.module.less
@@ -10,7 +10,7 @@
   margin: 10px;
 }
 
-.matrixPalette-row{
+.matrixPalette-row {
   display: flex;
   flex-direction: row;
 }
@@ -25,20 +25,4 @@
   display: inline-block;
   height: 20px;
   width: 1px;
-}
-
-.palette-view {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.palette-info {
-  font-size: 12px;
-  line-height: 16px;
-  width: 140px;
-
-  p {
-    margin: 0;
-  }
 }
