+++ demo/src/components/Swatch/colors.tsx
