+++ demo/src/components/Swatch/tooltip.tsx
