+++ demo/src/index.less
@@ -1,0 +1,15 @@
+.palette-view {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+}
+
+.palette-info {
+  font-size: 12px;
+  line-height: 16px;
+  width: 140px;
+
+  p {
+    margin: 0;
+  }
+}
