+++ package.json
@@ -51,14 +51,15 @@
   "bugs": {
     "url": "https://github.com/antvis/color-palette-json-schema/issues"
   },
-  "dependencies": {},
+  "dependencies": {
+    "@types/chroma-js": "^2.1.3"
+  },
   "devDependencies": {
     "@babel/runtime": "^7.13.10",
     "@commitlint/cli": "^12.1.4",
     "@commitlint/config-conventional": "^12.1.4",
     "@rollup/plugin-node-resolve": "^13.0.0",
     "@rollup/plugin-typescript": "^8.2.1",
-    "@types/chroma-js": "^2.1.3",
     "@types/jest": "^26.0.22",
     "@types/jest-json-schema": "^2.1.3",
     "@typescript-eslint/eslint-plugin": "^4.28.3",
