+++ src/index.ts
@@ -1,4 +1,4 @@
-import pkg from "../package.json";
+import pkg from '../package.json';
 export const version = pkg.version;
 
-export * from "./schema";
+export * from './schema';
